# swagger.json
{
    "openapi": "3.0.0",
    "info": {
        "version": "2.0.0",
        "title": "Polanco.dev API",
        "description": "API RESTful que alimenta Polanco.dev — blog técnico, portafolio profesional, cheatsheets y trucos rápidos. Construida con Node.js, Express y MongoDB."
    },
    "servers": [
        {
            "url": "http://localhost:3000/api/v1",
            "description": "Servidor de desarrollo local"
        },
        {
            "url": "https://your-app.up.railway.app/api/v1",
            "description": "Servidor de producción (Railway)"
        }
    ],
...
    "paths": {
...
        "/articles": {
            "get": {
                "tags": ["Articles"],
                "summary": "Listar artículos",
                "description": "Devuelve todos los artículos publicados. Soporta filtrado, ordenamiento, paginación y selección de campos.",
                "parameters": [
                    { "name": "category", "in": "query", "schema": { "type": "string", "enum": ["Programacion", "Idioma"] }, "description": "Filtrar por categoría" },
                    { "name": "sort", "in": "query", "schema": { "type": "string" }, "description": "Ordenar por campo(s), ej: -createdAt,title" },
                    { "name": "fields", "in": "query", "schema": { "type": "string" }, "description": "Seleccionar campos, ej: title,description" },
                    { "name": "page", "in": "query", "schema": { "type": "integer", "default": 1 } },
                    { "name": "limit", "in": "query", "schema": { "type": "integer", "default": 10 } },
                    { "name": "tags", "in": "query", "schema": { "type": "string" }, "description": "Filtrar por tags, ej: Angular,TypeScript" }
                ],
                "responses": {
                    "200": { "description": "Ok", "content": { "application/json": { "schema": { "type": "array", "items": { "$ref": "#/components/schemas/Article" } } } } }
                }
            },
            "post": {
                "tags": ["Articles"],
                "summary": "Crear artículo",
                "description": "Crea un nuevo artículo. Solo admin.",
                "security": [{ "bearerAuth": [] }],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": { "$ref": "#/components/schemas/ArticleInput" }
                        }
                    }
                },
                "responses": {
                    "201": { "description": "Artículo creado", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Article" } } } },
                    "400": { "$ref": "#/components/responses/400" },
                    "401": { "$ref": "#/components/responses/401" },
                    "403": { "$ref": "#/components/responses/403" }
                }
            }
        },
...
        "/articles/search": {
            "get": {
                "tags": ["Articles"],
                "summary": "Buscar artículos",
                "description": "Búsqueda de texto completo en título, descripción y contenido.",
                "parameters": [
                    { "name": "q", "in": "query", "required": true, "schema": { "type": "string" }, "description": "Término de búsqueda", "example": "typescript" }
                ],
                "responses": {
                    "200": { "description": "Resultados de búsqueda", "content": { "application/json": { "schema": { "type": "array", "items": { "$ref": "#/components/schemas/Article" } } } } }
                }
            }
        },
        "/articles/drafts": {
            "get": {
                "tags": ["Articles"],
                "summary": "Listar borradores",
                "description": "Devuelve artículos en estado borrador. Solo admin.",
                "security": [{ "bearerAuth": [] }],
                "responses": {
                    "200": { "description": "Ok" },
                    "401": { "$ref": "#/components/responses/401" },
                    "403": { "$ref": "#/components/responses/403" }
                }
            }
        },
        "/articles/admin/stats": {
            "get": {
                "tags": ["Articles"],
                "summary": "Estadísticas del dashboard",
                "description": "Devuelve estadísticas agregadas para el panel de administración (total artículos, vistas, likes, etc.). Solo admin.",
                "security": [{ "bearerAuth": [] }],
                "responses": {
                    "200": { "description": "Ok" },
                    "401": { "$ref": "#/components/responses/401" },
                    "403": { "$ref": "#/components/responses/403" }
                }
            }
        },
        "/articles/{id}": {
            "get": {
                "tags": ["Articles"],
                "summary": "Obtener artículo por slug",
                "description": "Busca el artículo por su slug único, generado automáticamente desde el título al crear o actualizar el artículo.",
                "parameters": [
                    { "name": "id", "in": "path", "required": true, "schema": { "type": "string" }, "description": "Slug del artículo (generado desde el título)", "example": "guia-completa-de-typescript" }
                ],
                "responses": {
                    "200": { "description": "Ok", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Article" } } } },
                    "404": { "$ref": "#/components/responses/404" }
                }
            },
            "patch": {
                "tags": ["Articles"],
                "summary": "Actualizar artículo",
                "description": "Actualiza parcialmente un artículo. Solo admin.",
                "security": [{ "bearerAuth": [] }],
                "parameters": [
                    { "name": "id", "in": "path", "required": true, "schema": { "type": "string" } }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": { "$ref": "#/components/schemas/ArticleInput" }
                        }
                    }
                },
                "responses": {
                    "200": { "description": "Artículo actualizado", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Article" } } } },
                    "400": { "$ref": "#/components/responses/400" },
                    "401": { "$ref": "#/components/responses/401" },
                    "403": { "$ref": "#/components/responses/403" },
                    "404": { "$ref": "#/components/responses/404" }
                }
            },
            "delete": {
                "tags": ["Articles"],
                "summary": "Eliminar artículo",
                "description": "Elimina un artículo por ID. Solo admin.",
                "security": [{ "bearerAuth": [] }],
                "parameters": [
                    { "name": "id", "in": "path", "required": true, "schema": { "type": "string" } }
                ],
                "responses": {
                    "204": { "description": "Artículo eliminado" },
                    "401": { "$ref": "#/components/responses/401" },
                    "403": { "$ref": "#/components/responses/403" },
                    "404": { "$ref": "#/components/responses/404" }
                }
            }
        },
        "/articles/{id}/view": {
            "patch": {
                "tags": ["Articles"],
                "summary": "Incrementar contador de vistas",
                "description": "Incrementa en 1 el contador de vistas del artículo. Requiere autenticación con rol admin. Rate limited: 1 por IP/slug cada 10 minutos.",
                "security": [{ "bearerAuth": [] }],
                "parameters": [
                    { "name": "id", "in": "path", "required": true, "schema": { "type": "string" }, "description": "Slug del artículo" }
                ],
                "responses": {
                    "200": { "description": "Vistas incrementadas" },
                    "401": { "$ref": "#/components/responses/401" },
                    "403": { "$ref": "#/components/responses/403" },
                    "404": { "$ref": "#/components/responses/404" }
                }
            }
        },
        "/articles/{id}/like": {
            "patch": {
                "tags": ["Articles"],
                "summary": "Dar like a un artículo",
                "description": "Incrementa en 1 el contador de likes del artículo. Requiere autenticación con rol admin. Rate limited: 1 por IP/slug cada 24 horas.",
                "security": [{ "bearerAuth": [] }],
                "parameters": [
                    { "name": "id", "in": "path", "required": true, "schema": { "type": "string" }, "description": "Slug del artículo" }
                ],
                "responses": {
                    "200": { "description": "Like registrado" },
                    "401": { "$ref": "#/components/responses/401" },
                    "403": { "$ref": "#/components/responses/403" },
                    "404": { "$ref": "#/components/responses/404" }
                }
            }
        },
        "/articles/{id}/related": {
            "get": {
                "tags": ["Articles"],
                "summary": "Artículos relacionados",
                "description": "Devuelve hasta 4 artículos de la misma categoría, excluyendo el actual. Recibe el slug del artículo.",
                "parameters": [
                    { "name": "id", "in": "path", "required": true, "schema": { "type": "string" }, "description": "MongoDB ObjectId o slug del artículo" }
                ],
                "responses": {
                    "200": { "description": "Ok", "content": { "application/json": { "schema": { "type": "array", "items": { "$ref": "#/components/schemas/Article" } } } } },
                    "404": { "$ref": "#/components/responses/404" }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
            "User": {
                "type": "object",
                "properties": {
                    "_id": { "type": "string" },
                    "name": { "type": "string" },
                    "email": { "type": "string", "format": "email" },
                    "photo": { "type": "string" },
                    "role": { "type": "string", "enum": ["user", "admin"] },
                    "active": { "type": "boolean" },
                    "passwordChangedAt": { "type": "string", "format": "date-time" }
                }
            },
...
            "Article": {
                "type": "object",
                "properties": {
                    "_id": { "type": "string", "example": "615930d4f37971021d64b413" },
                    "slug": { "type": "string", "description": "Slug único generado automáticamente desde el título (solo minúsculas, separado por guiones)", "example": "guia-completa-de-typescript", "readOnly": true },
                    "title": { "type": "string", "example": "Guía completa de TypeScript" },
                    "description": { "type": "string" },
                    "content": { "type": "string" },
                    "imageCover": { "type": "string" },
                    "category": { "type": "string", "enum": ["Programacion", "Idioma"] },
                    "tags": { "type": "array", "items": { "type": "string" } },
                    "author": { "$ref": "#/components/schemas/User" },
                    "views": { "type": "integer", "default": 0 },
                    "likes": { "type": "integer", "default": 0 },
                    "published": { "type": "boolean", "default": true, "description": "false = borrador, true = publicado" },
                    "createdAt": { "type": "string", "format": "date-time" }
                }
            },
            "ArticleInput": {
                "required": ["title", "description", "category"],
                "properties": {
                    "title": { "type": "string", "minLength": 10, "maxLength": 40, "example": "Guía completa de TypeScript", "description": "Título único. El slug se genera automáticamente." },
                    "description": { "type": "string", "example": "Aprende TypeScript desde cero" },
                    "content": { "type": "string", "description": "Contenido en Markdown" },
                    "imageCover": { "type": "string", "description": "URL de la imagen de portada (obtenida de POST /upload/article-image)" },
                    "category": { "type": "string", "enum": ["Programacion", "Idioma"] },
                    "tags": { "type": "array", "items": { "type": "string" } },
                    "published": { "type": "boolean", "default": true, "description": "false para guardar como borrador" }
                }
            },
...
        },
        "responses": {
            "400": { "description": "Solicitud incorrecta — parámetros inválidos o faltantes." },
            "401": { "description": "No autorizado — token JWT faltante o inválido." },
            "403": { "description": "Prohibido — no tienes permisos para realizar esta acción." },
            "404": { "description": "No encontrado — el recurso solicitado no existe." }
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "scheme": "bearer",
                "bearerFormat": "JWT",
                "description": "Token JWT obtenido en /users/login. Formato: Bearer <token>"
            }
        }
    }
}